[CCUBE-1510-2][JY] Shift resetting of form to useEffect in basic details, Add logic in mandatory field to repopulate default values.
