+++ src/components/element-editor/basic-details/common/type-field.tsx
@@ -4,49 +4,23 @@ import {
     useFormContext,
 } from "react-hook-form";
 import { IconDropdown } from "src/components/common/icon-dropdown";
-import { EElementType, TElement, useBuilder } from "src/context-providers";
+import { EElementType, useBuilder } from "src/context-providers";
 import { TFormFieldValues } from "src/schemas";
-import { ElementObjectGenerator } from "src/util";
 
 export const TypeField = () => {
     // =========================================================================
     // CONST, STATE, REF
     // =========================================================================
 
-    const { selectElementType, orderedIdentifiers, elements, focusedElement } =
-        useBuilder();
+    const { selectElementType } = useBuilder();
     const {
         control,
         formState: { errors },
-        reset,
     } = useFormContext<TFormFieldValues>();
 
     // =========================================================================
     // HELPER FUNCTIONS
     // =========================================================================
-
-    const resetFieldValues = (type: EElementType) => {
-        const existingIdentifiers = orderedIdentifiers
-            .filter(
-                (elementId) =>
-                    elementId.internalId !== focusedElement.element.internalId
-            )
-            .map((elementId) => elementId.internalId);
-
-        const existingIds = Object.values(elements)
-            .filter((element) => element.id !== focusedElement.element.id)
-            .map((element) => element.id);
-
-        const newElement: TElement = ElementObjectGenerator.generate(
-            type,
-            existingIdentifiers,
-            existingIds
-        );
-
-        reset(newElement, {
-            keepDefaultValues: true, // Dirty form upon changing type
-        });
-    };
 
     const handleChange = (
         value: EElementType,
@@ -56,10 +30,6 @@ export const TypeField = () => {
             return;
         }
         selectElementType(value);
-
-        // Reset all field values and populate with default values
-        resetFieldValues(value);
-
         field.onChange(value);
     };
